+++ src/other/utils.ts
@@ -1,0 +1,8 @@
+import * as path from 'path';
+import * as fs from 'fs';
+
+export function ensureDir(path: string) {
+  if (!fs.existsSync(path)) {
+    fs.mkdirSync(path, { recursive: true });
+  }
+}